requirements.txt
streamlit
groq
python-dotenv
pandas
PyPDF2
